# docs/swagger.yaml
definitions:
  handlers.JWTOutput:
    properties:
      expires:
        type: string
      token:
        type: string
    type: object
  models.Recipe:
    properties:
      _id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      publishedAt:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.User:
    properties:
      passsword:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /recipes:
    get:
      consumes:
      - application/json
      description: get all recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List recipes
      tags:
      - recipes
    post:
      consumes:
      - application/json
      description: Add new recipe
      parameters:
      - description: recipe object
        in: body
        name: recipe
        schema:
          $ref: '#/definitions/models.Recipe'
      - description: JWT Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add new recipe
      tags:
      - recipes
  /recipes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete recipe
      parameters:
      - description: recipe id
        in: path
        name: id
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete recipe
      tags:
      - recipes
    put:
      consumes:
      - application/json
      description: Update recipe
      parameters:
      - description: recipe id
        in: path
        name: id
        type: string
      - description: recipe object
        in: body
        name: recipe
        schema:
          $ref: '#/definitions/models.Recipe'
      - description: JWT Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update recipe
      tags:
      - recipes
  /recipes/search:
    get:
      consumes:
      - application/json
      description: get all recipes by tag
      parameters:
      - description: recipe search by tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List recipes by tag
      tags:
      - recipes
  /singin:
    post:
      consumes:
      - application/json
      description: Sigin API with username and password
      parameters:
      - description: user object
        in: body
        name: user
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JWTOutput'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Sigin API with username and password
      tags:
      - recipes
swagger: "2.0"